+++ test/testshortestpath.jl
@@ -2,7 +2,8 @@
 
     @testset "Simple Problem 1" begin
 
-        conns = [
+        problem = ShortestPathProblem(
+            Connection[
             Connection(1, 2, 3),
             Connection(1, 3, 2),
             Connection(1, 4, 4),
@@ -12,11 +13,12 @@
             Connection(4, 6, 2),
             Connection(5, 7, 6),
             Connection(6, 7, 5),
-        ]
-        # Shortest path: 1 -> 3 -> 6 -> 7
-        CorrectShortestPath = [conns[2], conns[6], conns[9]]
+        ])
 
-        result = solve(conns, problem = ShortestPathProblem)
+        # Shortest path: 1 -> 3 -> 6 -> 7
+        CorrectShortestPath = [problem.connections[2], problem.connections[6], problem.connections[9]]
+
+        result = solve(problem)
 
         @test result isa ShortestPathResult
         @test result.cost == 8.0